apps/backend/service/pyproject.toml
[tool.poetry]
name = "backend-fastapi"
version = "0.1.0"
description = ""
authors = ["Vasily Kovalev <minskij10@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
# Base
fastapi = {extras = ["full"], version = "^0.111.1"}
strawberry-graphql = { version = "^0.262.5", extras = ["fastapi"] } 
pydantic-extra-types = "^2.10.3"

user-agents= "^2.2.0"
pycountry = "^24.6.1"
jsonschema ="^4.23.0"
thefuzz ="0.22.1"
# Low level
ujson = "^5.10.0"
#starlette = "^0.37.2"
# Deploy
uvicorn = "^0.30.4"
gunicorn = "^20.0.4"
# Other
pydantic = {extras = ["email"], version = "^2.8.2"}
pydantic-settings = "^2.4.0"
python-dotenv = "^1.0.1"
# Common
cachetools = "^4.1.0"
tenacity = "^6.2.0"

PyJWT= "2.9.0"
passlib = "1.7.4"

# Postgres
SQLAlchemy = "1.4.46"
SQLAlchemy-serializer = "1.4.1"
psycopg2-binary = "2.9.9"
alembic = "1.9.4"

kagglehub = "0.3.4"

[tool.poetry.dev-dependencies]
# Lint
black = "25.1.0"
pylint = "^3.3.6"
isort = "^4.3"
unify = "^0.5.0"
# Typings
mypy = "1.15.0"
# Test
pytest-mock = "^1.13"

pytest-dotenv = "^0.5.2"
pytest = "^6.2.5"
pytest-asyncio = "^0.15.1"
faker = "^26.1.0"
python-semantic-release = "^9.8.5"
pre-commit = "^3.8.0"
pytest-cov = "^5.0.0"
anyio = "^4.4.0"
typing-extensions = "^4.12.2"

[build-system]
requires = ["poetry>=1.8.3"]
build-backend = "poetry.masonry.api"
